Changing navigate to Pocket URL modifier on Windows to Alt key
