+++ src/Types.ts
@@ -1,0 +1,1 @@
+export type SupportedPlatform = "mac" | "windows" | "linux";
